fix training etc...
