+++ monailabel/utils/datastore/dicom/util.py
@@ -8,130 +8,22 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-import json
 import logging
 import os
-import pathlib
-import tempfile
 import time
 from hashlib import md5
 
-import numpy as np
-import pydicom_seg
-import SimpleITK
 from dicomweb_client import DICOMwebClient
-from monai.data import write_nifti
-from monai.transforms import LoadImage
+from dicomweb_client.api import load_json_dataset
 from pydicom.filereader import dcmread
 
-from monailabel.utils.datastore.dicom.colors import GENERIC_ANATOMY_COLORS
 from monailabel.utils.others.generic import run_command
-from monailabel.utils.others.writer import write_itk
 
 logger = logging.getLogger(__name__)
 
 
 def generate_key(patient_id: str, study_id: str, series_id: str):
     return md5(f"{patient_id}+{study_id}+{series_id}".encode("utf-8")).hexdigest()
-
-
-def binary_to_image(reference_image, label, dtype=np.uint16, file_ext=".nii.gz", use_itk=True):
-    image_np, meta_dict = LoadImage()(reference_image)
-    label_np = np.fromfile(label, dtype=dtype)
-
-    logger.info(f"Image: {image_np.shape}")
-    logger.info(f"Label: {label_np.shape}")
-
-    label_np = label_np.reshape(image_np.shape, order="F")
-    logger.info(f"Label (reshape): {label_np.shape}")
-
-    output_file = tempfile.NamedTemporaryFile(suffix=file_ext).name
-    affine = meta_dict.get("affine")
-    if use_itk:
-        write_itk(label_np, output_file, affine=affine, dtype=None, compress=True)
-    else:
-        write_nifti(label_np, output_file, affine=affine)
-    return output_file
-
-
-def nifti_to_dicom_seg(series_dir, label, label_info, file_ext="*"):
-    label_np, meta_dict = LoadImage()(label)
-    unique_labels = np.unique(label_np.flatten()).astype(np.int)
-    unique_labels = unique_labels[unique_labels != 0]
-
-    segment_attributes = []
-    for i, idx in enumerate(unique_labels):
-        info = label_info[i] if label_info and i < len(label_info) else {}
-        name = info.get("name", "unknown")
-        description = info.get("description", "Unknown")
-        rgb = list(info.get("color", GENERIC_ANATOMY_COLORS.get(name, (255, 0, 0))))[0:3]
-        rgb = [int(x) for x in rgb]
-
-        logger.info(f"{i} => {idx} => {name}")
-
-        segment_attribute = info.get(
-            "segmentAttribute",
-            {
-                "labelID": int(idx),
-                "SegmentLabel": name,
-                "SegmentDescription": description,
-                "SegmentAlgorithmType": "AUTOMATIC",
-                "SegmentAlgorithmName": "MONAILABEL",
-                "SegmentedPropertyCategoryCodeSequence": {
-                    "CodeValue": "123037004",
-                    "CodingSchemeDesignator": "SCT",
-                    "CodeMeaning": "Anatomical Structure",
-                },
-                "SegmentedPropertyTypeCodeSequence": {
-                    "CodeValue": "78961009",
-                    "CodingSchemeDesignator": "SCT",
-                    "CodeMeaning": name,
-                },
-                "recommendedDisplayRGBValue": rgb,
-            },
-        )
-        segment_attributes.append(segment_attribute)
-
-    template = {
-        "ContentCreatorName": "Reader1",
-        "ClinicalTrialSeriesID": "Session1",
-        "ClinicalTrialTimePointID": "1",
-        "SeriesDescription": "Segmentation",
-        "SeriesNumber": "300",
-        "InstanceNumber": "1",
-        "segmentAttributes": [segment_attributes],
-        "ContentLabel": "SEGMENTATION",
-        "ContentDescription": "MONAI Label - Image segmentation",
-        "ClinicalTrialCoordinatingCenterName": "MONAI",
-        "BodyPartExamined": "",
-    }
-
-    logger.info(json.dumps(template, indent=2))
-    if not segment_attributes:
-        logger.error("Missing Attributes/Empty Label provided")
-        return None
-
-    template = pydicom_seg.template.from_dcmqi_metainfo(template)
-    writer = pydicom_seg.MultiClassWriter(
-        template=template,
-        inplane_cropping=False,
-        skip_empty_slices=True,
-        skip_missing_segment=False,
-    )
-
-    # Read source Images
-    series_dir = pathlib.Path(series_dir)
-    image_files = series_dir.glob(file_ext)
-    image_datasets = [dcmread(str(f), stop_before_pixels=True) for f in image_files]
-    logger.info(f"Total Source Images: {len(image_datasets)}")
-
-    mask = SimpleITK.ReadImage(label)
-    mask = SimpleITK.Cast(mask, SimpleITK.sitkUInt16)
-
-    output_file = tempfile.NamedTemporaryFile(suffix=".dcm").name
-    dcm = writer.write(mask, image_datasets)
-    dcm.save_as(output_file)
-    return output_file
 
 
 def get_scu(query, output_dir, query_level="SERIES", host="127.0.0.1", port="4242", aet="MONAILABEL"):
@@ -165,27 +57,42 @@ def score_scu(input, host="127.0.0.1", port="4242", aet="MONAILABEL"):
     inputs = input if isinstance(input, list) else [input]
     for i in inputs:
         run_command("python", ["-m", "pynetdicom", "storescu", host, port, "-aet", aet, i])
-    logger.info(f"Time to run GET-SCU: {time.time() - start} (sec)")
+    logger.info(f"Time to run STORE-SCU: {time.time() - start} (sec)")
 
 
-def dicom_web_download_series(study_id, series_id, save_dir, url="http://127.0.0.1:8042/dicom-web"):
+def dicom_web_download_series(study_id, series_id, save_dir, client: DICOMwebClient):
     start = time.time()
 
+    # Limitation for DICOMWeb Client as it needs StudyInstanceUID to fetch series
+    if not study_id:
+        meta = load_json_dataset(client.search_for_series(search_filters={"SeriesInstanceUID": series_id})[0])
+        study_id = str(meta["StudyInstanceUID"].value)
+
     os.makedirs(save_dir, exist_ok=True)
-    client = DICOMwebClient(url=url)
     instances = client.retrieve_series(study_id, series_id)
     for instance in instances:
-        instance_id = instance.SOPInstanceUID
+        instance_id = instance["SOPInstanceUID"]
         file_name = os.path.join(save_dir, f"{instance_id}.dcm")
         instance.save_as(file_name)
 
     logger.info(f"Time to download: {time.time() - start} (sec)")
 
 
-def dicom_web_upload_dcm(input, url="http://127.0.0.1:8042/dicom-web"):
+def dicom_web_upload_dcm(input, client: DICOMwebClient):
     start = time.time()
     dataset = dcmread(input)
+    result = client.store_instances([dataset])
 
-    client = DICOMwebClient(url=url)
-    client.store_instances([dataset])
-    logger.info(f"Time to save: {time.time() - start} (sec)")
+    url = ""
+    for elm in result.iterall():
+        s = str(elm.value)
+        logger.info(f"{s}")
+        if "/series/" in s:
+            url = s
+            break
+
+    series_id = url.split("/series/")[1].split("/")[0] if url else ""
+    logger.info(f"Series Instance UID: {series_id}")
+
+    logger.info(f"Time to upload: {time.time() - start} (sec)")
+    return series_id
