+++ monailabel/utils/datastore/local.py
@@ -8,7 +8,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
+import copy
 import fnmatch
 import io
 import json
@@ -26,31 +26,87 @@ from watchdog.observers import Observer
 
 from monailabel.interfaces.datastore import Datastore, DefaultLabelTag
 from monailabel.interfaces.exception import ImageNotFoundException, LabelNotFoundException
-from monailabel.utils.others.generic import file_checksum
+from monailabel.utils.others.generic import file_checksum, remove_file
 
 logger = logging.getLogger(__name__)
 
 
-class ImageModel(BaseModel):
+class DataModel(BaseModel):
     id: str
+    ext: str = ""
     info: Dict[str, Any] = {}
 
-
-class LabelModel(BaseModel):
-    id: str
-    tag: str
-    info: Dict[str, Any] = {}
+    def path(self):
+        return self.id + self.ext
 
 
-class ObjectModel(BaseModel):
-    image: ImageModel
-    labels: List[LabelModel] = []
+class ImageLabelModel(BaseModel):
+    image: DataModel
+    labels: Dict[str, DataModel] = {}  # tag => label
+
+    def label(self, id):
+        for tag, label in self.labels.items():
+            if label.id == id:
+                return tag, label
+        return None, None
+
+    def tags(self):
+        return self.labels.keys()
 
 
 class LocalDatastoreModel(BaseModel):
     name: str
     description: str
-    objects: List[ObjectModel] = []
+    images_dir: str = "."
+    labels_dir: str = "labels"
+    objects: Dict[str, ImageLabelModel] = {}
+
+    # will be ignored while saving...
+    base_path: str = ""
+
+    def tags(self):
+        tags = set()
+        for v in self.objects.values():
+            tags.update(v.tags())
+        return tags
+
+    def filter_by_tag(self, tag: str):
+        return [obj for obj in self.objects.values() if obj.labels.get(tag)]
+
+    def filter_by_id(self, id: str):
+        return [obj for obj in self.objects.values() if obj.label(id)[0]]
+
+    def label(self, id: str):
+        objects = self.filter_by_id(id)
+        obj = next(iter(objects)) if objects else None
+        if obj:
+            return obj.label(id)
+        return None, None
+
+    def image_path(self):
+        return os.path.join(self.base_path, self.images_dir) if self.base_path else self.images_dir
+
+    def label_path(self, tag):
+        path = os.path.join(self.labels_dir, tag) if tag else self.labels_dir
+        return os.path.join(self.base_path, path) if self.base_path else path
+
+    def labels_path(self):
+        path = self.labels_dir
+        return {tag: os.path.join(path, tag) if self.base_path else path for tag in self.tags()}
+
+    def datalist(self, tag: str) -> List[Dict[str, str]]:
+        image_path = self.image_path()
+        label_path = self.label_path(tag)
+
+        items = []
+        for obj in self.filter_by_tag(tag):
+            items.append(
+                {
+                    "image": os.path.realpath(os.path.join(image_path, obj.image.path())),
+                    "label": os.path.realpath(os.path.join(label_path, obj.labels[tag].path())),
+                }
+            )
+        return items
 
 
 class LocalDatastore(Datastore):
@@ -69,10 +125,10 @@ class LocalDatastore(Datastore):
     def __init__(
         self,
         datastore_path: str,
-        datastore_config: str = "datastore.json",
-        label_store_path: str = "labels",
-        image_extensions=("*.nii.gz", "*.nii"),
-        label_extensions=("*.nii.gz", "*.nii"),
+        images_dir: str = ".",
+        labels_dir: str = "labels",
+        datastore_config: str = "datastore_v2.json",
+        extensions=("*.nii.gz", "*.nii"),
         auto_reload=False,
     ):
         """
@@ -88,41 +144,47 @@ class LocalDatastore(Datastore):
         """
         self._datastore_path = datastore_path
         self._datastore_config_path = os.path.join(datastore_path, datastore_config)
-        self._label_store_path = label_store_path
-        self._image_extensions = [image_extensions] if isinstance(image_extensions, str) else image_extensions
-        self._label_extensions = [label_extensions] if isinstance(label_extensions, str) else label_extensions
+        self._extensions = [extensions] if isinstance(extensions, str) else extensions
         self._ignore_event_count = 0
         self._ignore_event_config = False
         self._config_ts = 0
         self._auto_reload = auto_reload
 
-        self._lock = FileLock(os.path.join(datastore_path, ".lock"))
         logging.getLogger("filelock").setLevel(logging.ERROR)
 
-        logger.info(f"Image Extensions: {self._image_extensions}")
-        logger.info(f"Label Extensions: {self._label_extensions}")
+        logger.info(f"Extensions: {self._extensions}")
         logger.info(f"Auto Reload: {auto_reload}")
 
-        self._datastore: LocalDatastoreModel = LocalDatastoreModel(name="new-dataset", description="New Dataset")
+        os.makedirs(self._datastore_path, exist_ok=True)
+
+        self._lock = FileLock(os.path.join(datastore_path, ".lock"))
+        self._datastore: LocalDatastoreModel = LocalDatastoreModel(
+            name="new-dataset", description="New Dataset", images_dir=images_dir, labels_dir=labels_dir
+        )
+        self._datastore.base_path = self._datastore_path
         self._init_from_datastore_file(throw_exception=True)
 
-        # ensure labels path exists regardless of whether a datastore file is present
-        os.makedirs(os.path.join(self._datastore_path, self._label_store_path), exist_ok=True)
+        os.makedirs(self._datastore.image_path(), exist_ok=True)
+        os.makedirs(self._datastore.label_path(None), exist_ok=True)
+        os.makedirs(self._datastore.label_path(DefaultLabelTag.FINAL), exist_ok=True)
+        os.makedirs(self._datastore.label_path(DefaultLabelTag.ORIGINAL), exist_ok=True)
 
         # reconcile the loaded datastore file with any existing files in the path
         self._reconcile_datastore()
 
         if auto_reload:
             logger.info("Start observing external modifications on datastore (AUTO RELOAD)")
-            include_patterns = [
-                f"{self._datastore_path}{os.path.sep}{ext}" for ext in [*image_extensions, *label_extensions]
-            ]
-            include_patterns.extend(
-                [
-                    f"{os.path.join(self._datastore_path, self._label_store_path)}{os.path.sep}{ext}"
-                    for ext in [*image_extensions, *label_extensions]
-                ]
-            )
+            # Image Dir
+            include_patterns = [f"{self._datastore.image_path()}{os.path.sep}{ext}" for ext in [*extensions]]
+
+            # Label Dir(s)
+            label_dirs = self._datastore.labels_path()
+            label_dirs[DefaultLabelTag.FINAL] = self._datastore.label_path(DefaultLabelTag.FINAL)
+            label_dirs[DefaultLabelTag.ORIGINAL] = self._datastore.label_path(DefaultLabelTag.ORIGINAL)
+            for label_dir in label_dirs.values():
+                include_patterns.extend(f"{label_dir}{os.path.sep}{ext}" for ext in [*extensions])
+
+            # Config
             include_patterns.append(self._datastore_config_path)
 
             self._handler = PatternMatchingEventHandler(patterns=include_patterns)
@@ -179,14 +241,6 @@ class LocalDatastore(Datastore):
         self._datastore.description = description
         self._update_datastore_file()
 
-    @property
-    def datastore_path(self):
-        return self._datastore_path
-
-    @property
-    def labelstore_path(self):
-        return os.path.join(self._datastore_path, self._label_store_path)
-
     def datalist(self, full_path=True) -> List[Dict[str, str]]:
         """
         Return a dictionary of image and label pairs corresponding to the 'image' and 'label'
@@ -194,18 +248,13 @@ class LocalDatastore(Datastore):
 
         :return: the {'label': image, 'label': label} pairs for training
         """
-        items = []
-        for obj in self._datastore.objects:
-            image_path = self._get_path(obj.image.id, False, full_path)
-            for label in obj.labels:
-                if label.tag == DefaultLabelTag.FINAL:
-                    items.append(
-                        {
-                            "image": image_path,
-                            "label": self._get_path(label.id, True, full_path),
-                        }
-                    )
-        return items
+        ds = self._datastore.datalist(DefaultLabelTag.FINAL)
+        if not full_path:
+            ds = json.loads(json.dumps(ds).replace(f"{self._datastore_path.rstrip(os.pathsep)}{os.pathsep}", ""))
+        return ds
+
+    def to_bytes(self, file):
+        return io.BytesIO(pathlib.Path(file).read_bytes())
 
     def get_image(self, image_id: str) -> Any:
         """
@@ -214,12 +263,12 @@ class LocalDatastore(Datastore):
         :param image_id: the desired image's id
         :return: return the "image"
         """
-        buf = None
-        for obj in self._datastore.objects:
-            if obj.image.id == image_id:
-                buf = io.BytesIO(pathlib.Path(os.path.join(self._datastore_path, obj.image.id)).read_bytes())
-                break
-        return buf
+        obj = self._datastore.objects.get(image_id)
+        return (
+            self.to_bytes(os.path.realpath(os.path.join(self._datastore.image_path(), obj.image.path())))
+            if obj
+            else None
+        )
 
     def get_image_uri(self, image_id: str) -> str:
         """
@@ -228,12 +277,8 @@ class LocalDatastore(Datastore):
         :param image_id: the desired image's id
         :return: return the image uri
         """
-        image_path = ""
-        for obj in self._datastore.objects:
-            if obj.image.id == image_id:
-                image_path = os.path.join(self._datastore_path, obj.image.id)
-                break
-        return image_path
+        obj = self._datastore.objects.get(image_id)
+        return str(os.path.realpath(os.path.join(self._datastore.image_path(), obj.image.path()))) if obj else ""
 
     def get_image_info(self, image_id: str) -> Dict[str, Any]:
         """
@@ -242,20 +287,17 @@ class LocalDatastore(Datastore):
         :param image_id: the desired image id
         :return: image info as a list of dictionaries Dict[str, Any]
         """
-        info = {}
-        for obj in self._datastore.objects:
-            if obj.image.id == image_id:
-                info = obj.image.info
-
-        local_path = pathlib.Path(os.path.join(self._datastore_path, image_id))
-        info.update(
-            {
-                "checksum": file_checksum(local_path),
-                "name": obj.image.id,
-                "path": local_path,
-            }
-        )
-
+        obj = self._datastore.objects.get(image_id)
+        info = copy.deepcopy(obj.image.info) if obj else {}
+        if obj:
+            path = os.path.realpath(os.path.join(self._datastore.image_path(), obj.image.path()))
+            info.update(
+                {
+                    "checksum": file_checksum(path),
+                    "name": obj.image.path(),
+                    "path": path,
+                }
+            )
         return info
 
     def get_label(self, label_id: str) -> Any:
@@ -265,14 +307,12 @@ class LocalDatastore(Datastore):
         :param label_id: the desired label's id
         :return: return the "label"
         """
-        buf = None
-        for obj in self._datastore.objects:
-            for label in obj.labels:
-                if label.id == label_id:
-                    buf = io.BytesIO(
-                        pathlib.Path(os.path.join(self._datastore_path, self._label_store_path, label.id)).read_bytes()
-                    )
-        return buf
+        tag, label = self._datastore.label(label_id)
+        return (
+            self.to_bytes(os.path.realpath(os.path.join(self._datastore.label_path(tag), label.path())))
+            if tag
+            else None
+        )
 
     def get_label_uri(self, label_id: str) -> str:
         """
@@ -281,11 +321,8 @@ class LocalDatastore(Datastore):
         :param label_id: the desired label's id
         :return: return the label uri
         """
-        for obj in self._datastore.objects:
-            for label in obj.labels:
-                if label.id == label_id:
-                    return os.path.join(self._datastore_path, self._label_store_path, label.id)
-        return ""
+        tag, label = self._datastore.label(label_id)
+        return str(os.path.realpath(os.path.join(self._datastore.label_path(tag), label.path()))) if tag else ""
 
     def get_labels_by_image_id(self, image_id: str) -> Dict[str, str]:
         """
@@ -294,9 +331,9 @@ class LocalDatastore(Datastore):
         :param image_id: the desired image's id
         :return: label ids mapped to the appropriate `LabelTag` as Dict[str, LabelTag]
         """
-        for obj in self._datastore.objects:
-            if obj.image.id == image_id:
-                return {label.id: label.tag for label in obj.labels}
+        obj = self._datastore.objects.get(image_id)
+        if obj:
+            return {label.id: tag for tag, label in obj.labels.items()}
         return {}
 
     def get_label_by_image_id(self, image_id: str, tag: str) -> str:
@@ -307,12 +344,9 @@ class LocalDatastore(Datastore):
         :param tag: matching tag name
         :return: label id
         """
-        for obj in self._datastore.objects:
-            if obj.image.id == image_id:
-                for label in obj.labels:
-                    if label.tag == tag:
-                        return label.id
-        return ""
+        obj = self._datastore.objects.get(image_id)
+        label = obj.labels.get(tag) if obj else None
+        return label.id if label else ""
 
     def get_label_info(self, label_id: str) -> Dict[str, Any]:
         """
@@ -321,12 +355,9 @@ class LocalDatastore(Datastore):
         :param label_id: the desired label id
         :return: label info as a list of dictionaries Dict[str, Any]
         """
-        for obj in self._datastore.objects:
-            for label in obj.labels:
-                if label.id == label_id:
-                    return label.info
-
-        return {}
+        _, label = self._datastore.label(label_id)
+        info: Dict[str, Any] = label.info if label else {}
+        return info
 
     def get_labeled_images(self) -> List[str]:
         """
@@ -334,13 +365,7 @@ class LocalDatastore(Datastore):
 
         :return: list of image ids List[str]
         """
-        image_ids = []
-        for obj in self._datastore.objects:
-            for label in obj.labels:
-                if label.tag == DefaultLabelTag.FINAL:
-                    image_ids.append(obj.image.id)
-
-        return image_ids
+        return [obj.image.id for obj in self._datastore.objects.values() if obj.labels.get(DefaultLabelTag.FINAL)]
 
     def get_unlabeled_images(self) -> List[str]:
         """
@@ -348,12 +373,7 @@ class LocalDatastore(Datastore):
 
         :return: list of image ids List[str]
         """
-        image_ids = []
-        for obj in self._datastore.objects:
-            if not obj.labels or DefaultLabelTag.FINAL not in [label.tag for label in obj.labels]:
-                image_ids.append(obj.image.id)
-
-        return image_ids
+        return [obj.image.id for obj in self._datastore.objects.values() if not obj.labels.get(DefaultLabelTag.FINAL)]
 
     def list_images(self) -> List[str]:
         """
@@ -361,7 +381,7 @@ class LocalDatastore(Datastore):
 
         :return: list of image ids List[str]
         """
-        return [obj.image.id for obj in self._datastore.objects]
+        return list(self._datastore.objects.keys())
 
     def _on_any_event(self, event):
         if self._ignore_event_count:
@@ -391,11 +411,17 @@ class LocalDatastore(Datastore):
         self._reconcile_datastore()
 
     def add_image(self, image_id: str, image_filename: str) -> str:
+        image_ext = "".join(pathlib.Path(image_filename).suffixes)
         if not image_id:
-            image_id = os.path.basename(image_filename)
+            image_id = os.path.basename(image_filename).replace(image_ext, "")
 
-        logger.info(f"Adding Image: {image_id}")
-        shutil.copy(image_filename, os.path.join(self._datastore_path, image_id))
+        logger.info(f"Adding Image: {image_id} => {image_filename}")
+        dest = os.path.realpath(os.path.join(self._datastore.image_path(), image_id + image_ext))
+        if os.path.isdir(image_filename):
+            shutil.copytree(image_filename, dest)
+        else:
+            shutil.copy(image_filename, dest)
+
         if not self._auto_reload:
             self.refresh()
         return image_id
@@ -409,63 +435,64 @@ class LocalDatastore(Datastore):
             self.remove_label(label_id)
 
         # Remove Image
-        p = os.path.join(self._datastore_path, image_id)
-        if os.path.exists(p):
-            os.unlink(p)
+        obj = self._datastore.objects.get(image_id)
+        p = os.path.realpath(os.path.join(self._datastore.image_path(), obj.image.path())) if obj else None
+        remove_file(p)
 
         if not self._auto_reload:
             self.refresh()
 
-    def save_label(self, image_id: str, label_filename: str, label_tag: str, label_info: Dict[str, Any]) -> str:
+    def save_label(
+        self, image_id: str, label_filename: str, label_tag: str, label_info: Dict[str, Any], label_id: str = ""
+    ) -> str:
         """
         Save a label for the given image id and return the newly saved label's id
 
         :param image_id: the image id for the label
         :param label_filename: the path to the label file
         :param label_tag: the tag for the label
+        :param label_info: additional info for the label
+        :param label_id: use this label id instead of generating it from filename
         :return: the label id for the given label filename
         """
         logger.info(f"Saving Label for Image: {image_id}; Tag: {label_tag}")
+        obj = self._datastore.objects.get(image_id)
+        if not obj:
+            raise ImageNotFoundException(f"Image {image_id} not found")
 
-        for obj in self._datastore.objects:
-            if obj.image.id == image_id:
-                image_ext = "".join(pathlib.Path(image_id).suffixes)
-                label_ext = "".join(pathlib.Path(label_filename).suffixes)
-                label_id = "label_" + label_tag + "_" + image_id.replace(image_ext, "") + label_ext
+        label_ext = "".join(pathlib.Path(label_filename).suffixes)
+        if not label_id:
+            label_id = image_id
+        else:
+            label_id = f"{image_id}+{label_id}"
 
-                with self._lock:
-                    os.makedirs(os.path.join(self._datastore_path, self._label_store_path), exist_ok=True)
-                    datastore_label_path = os.path.join(self._datastore_path, self._label_store_path, label_id)
-                    shutil.copy(src=label_filename, dst=datastore_label_path, follow_symlinks=True)
+        logger.info(f"Adding Label: {label_id} => {label_filename}")
+        label_path = self._datastore.label_path(label_tag)
+        dest = os.path.join(label_path, label_id + label_ext)
 
-                    lm = LabelModel(id=label_id, tag=label_tag, info={"ts": int(time.time())})
-                    if label_tag not in [label.tag for label in obj.labels]:
-                        obj.labels.append(lm)
-                    else:
-                        for label_index, label in enumerate(obj.labels):
-                            if label.tag == label_tag:
-                                obj.labels[label_index] = lm  # This will reset the previous info
+        with self._lock:
+            os.makedirs(label_path, exist_ok=True)
+            if os.path.isdir(label_filename):
+                shutil.copytree(label_filename, dest)
+            else:
+                shutil.copy(label_filename, dest)
 
-                    self._update_datastore_file(lock=False)
-                return label_id
-
-        raise ImageNotFoundException(f"Image {image_id} not found")
+            obj.labels[label_tag] = DataModel(id=label_id, info={"ts": int(time.time())}, ext=label_ext)
+            self._update_datastore_file(lock=False)
+        return label_id
 
     def remove_label(self, label_id: str) -> None:
         logger.info(f"Removing label: {label_id}")
-        p = os.path.join(self.labelstore_path, label_id)
-        if os.path.exists(p):
-            os.unlink(p)
+
+        tag, label = self._datastore.label(label_id)
+        p = os.path.realpath(os.path.join(self._datastore.label_path(tag), label.path())) if label else None
+        remove_file(p)
+
         if not self._auto_reload:
             self.refresh()
 
     def remove_label_by_tag(self, label_tag: str) -> None:
-        label_ids = []
-        for obj in self._datastore.objects:
-            for label in obj.labels:
-                if label.tag == label_tag:
-                    label_ids.append(label.id)
-
+        label_ids = [obj.labels[label_tag].id for obj in self._datastore.objects.values() if obj.labels.get(label_tag)]
         logger.info(f"Tag: {label_tag}; Removing label(s): {label_ids}")
         for label_id in label_ids:
             self.remove_label(label_id)
@@ -477,13 +504,12 @@ class LocalDatastore(Datastore):
         :param image_id: the id of the image we want to add/update info
         :param info: a dictionary of custom image information Dict[str, Any]
         """
-        for obj in self._datastore.objects:
-            if obj.image.id == image_id:
-                obj.image.info.update(info)
-                self._update_datastore_file()
-                return
+        obj = self._datastore.objects.get(image_id)
+        if not obj:
+            raise ImageNotFoundException(f"Image {image_id} not found")
 
-        raise ImageNotFoundException(f"Image {image_id} not found")
+        obj.image.info.update(info)
+        self._update_datastore_file()
 
     def update_label_info(self, label_id: str, info: Dict[str, Any]) -> None:
         """
@@ -492,26 +518,14 @@ class LocalDatastore(Datastore):
         :param label_id: the id of the label we want to add/update info
         :param info: a dictionary of custom label information Dict[str, Any]
         """
-        for obj in self._datastore.objects:
-            for label in obj.labels:
-                if label.id == label_id:
-                    label.info.update(info)
-                    self._update_datastore_file()
-                    return
+        _, label = self._datastore.label(label_id)
+        if not label:
+            raise LabelNotFoundException(f"Label {label_id} not found")
 
-        raise LabelNotFoundException(f"Label {label_id} not found")
+        label.info.update(info)
+        self._update_datastore_file()
 
-    def _get_path(self, path: str, is_label: bool, full_path=True):
-        if is_label:
-            path = os.path.join(self._label_store_path, path)
-
-        if not full_path or os.path.isabs(path):
-            return path
-
-        return os.path.realpath(os.path.join(self._datastore_path, path))
-
-    @staticmethod
-    def _list_files(path, patterns):
+    def _list_files(self, path, patterns):
         files = os.listdir(path)
 
         filtered = dict()
@@ -522,75 +536,90 @@ class LocalDatastore(Datastore):
         return filtered
 
     def _reconcile_datastore(self):
-        if self._remove_object_with_missing_file() or self._add_object_with_present_file():
+        invalidate = 0
+        invalidate += self._remove_non_existing()
+        invalidate += self._add_non_existing_images()
+
+        labels_dir = self._datastore.label_path(None)
+        logger.info(f"Labels Dir {labels_dir}")
+        tags = [f for f in os.listdir(labels_dir) if os.path.isdir(os.path.join(labels_dir, f))]
+        logger.debug(f"Label Tags: {tags}")
+        for tag in tags:
+            invalidate += self._add_non_existing_labels(tag)
+
+        invalidate += self._remove_non_existing()
+
+        logger.info(f"Invalidate count: {invalidate}")
+        if invalidate:
+            logger.debug("Save datastore file to disk")
             self._update_datastore_file()
+        else:
+            logger.debug("No changes needed to flush to disk")
 
-    def _remove_object_with_missing_file(self) -> bool:
-        """
-        remove objects present in the datastore file but not present on path
-        (even if labels exist, if images do not the whole object is removed from the datastore)
-        """
-        invalidate = False
+    def _add_non_existing_images(self) -> int:
+        invalidate = 0
 
-        image_id_files = self._list_files(self._datastore_path, self._image_extensions)
-        image_id_datastore = [obj.image.id for obj in self._datastore.objects]
-        missing_file_image_id = list(set(image_id_datastore) - set(image_id_files.keys()))
-        if missing_file_image_id:
-            logger.info(f"Removing Missing Images: {missing_file_image_id}")
-            invalidate = True
-            self._datastore.objects = [
-                obj for obj in self._datastore.objects if obj.image.id not in missing_file_image_id
-            ]
+        local_images = self._list_files(self._datastore.image_path(), self._extensions)
 
-        label_id_files = self._list_files(
-            os.path.join(self._datastore_path, self._label_store_path), self._label_extensions
-        )
-        label_id_datastore = [label.id for obj in self._datastore.objects for label in obj.labels]
-        missing_file_label_id = list(set(label_id_datastore) - set(label_id_files.keys()))
-        if missing_file_label_id:
-            logger.info(f"Removing Missing Labels: {missing_file_label_id}")
-            invalidate = True
-            for obj in self._datastore.objects:
-                obj.labels = [label for label in obj.labels if label.id not in missing_file_label_id]
+        image_ids = list(self._datastore.objects.keys())
+        for image_file in local_images:
+            image_ext = "".join(pathlib.Path(image_file).suffixes)
+            image_id = image_file.replace(image_ext, "")
+
+            if image_id not in image_ids:
+                logger.info(f"Adding New Image: {image_id} => {image_file}")
+                invalidate += 1
+                self._datastore.objects[image_id] = ImageLabelModel(image=DataModel(id=image_id, ext=image_ext))
 
         return invalidate
 
-    def _add_object_with_present_file(self) -> bool:
-        """
-        add objects which are not present in the datastore file, but are present in the datastore directory
-        this adds the image present in the datastore path and any corresponding labels for that image
-        """
-        invalidate = False
+    def _add_non_existing_labels(self, tag) -> int:
+        invalidate = 0
 
-        image_id_files = LocalDatastore._list_files(self._datastore_path, self._image_extensions)
-        label_id_files = LocalDatastore._list_files(
-            os.path.join(self._datastore_path, self._label_store_path), self._label_extensions
-        )
+        local_labels = self._list_files(self._datastore.label_path(tag), self._extensions)
 
-        # add any missing image files and any corresponding labels
-        existing_image_ids = [obj.image.id for obj in self._datastore.objects]
-        for image_id in image_id_files:
+        label_ids = [obj.labels[tag].id for obj in self._datastore.filter_by_tag(tag)]
+        for label_file in local_labels:
+            label_ext = "".join(pathlib.Path(label_file).suffixes)
+            label_id = label_file.replace(label_ext, "")
+            image_id = label_id.split("+")[0]
 
-            image_ext = "".join(pathlib.Path(image_id).suffixes)
-            image_id_nosuffix = image_id.replace(image_ext, "")
+            if label_id not in label_ids:
+                obj = self._datastore.objects.get(image_id)
+                if not obj:
+                    logger.warning(f"IGNORE:: No matching image '{image_id}' for '{label_id}' to add [{label_file}]")
+                    continue
 
-            # add the image i if not present
-            if image_id not in existing_image_ids:
-                logger.info(f"Adding New Image: {image_id}")
-                invalidate = True
-                self._datastore.objects.append(ObjectModel(image=ImageModel(id=image_id)))
+                logger.info(f"Adding New Label: {tag} => {label_id} => {label_file} for {image_id}")
+                obj.labels[tag] = DataModel(id=label_id, ext=label_ext)
+                invalidate += 1
 
-            # for matching image ids only
-            for label_id in [l_id for l_id in label_id_files if f"{image_id_nosuffix}." in l_id]:
-                image_id_index = [obj.image.id for obj in self._datastore.objects].index(image_id)
-                label_parts = label_id.split(image_id_nosuffix)
-                label_tag = label_parts[0].replace("label_", "").strip("_")
+        return invalidate
 
-                if label_id not in [label.id for label in self._datastore.objects[image_id_index].labels]:
-                    logger.info(f"Adding New Label: {image_id} => {label_id}")
-                    invalidate = True
-                    self._datastore.objects[image_id_index].labels.append(LabelModel(id=label_id, tag=label_tag))
+    def _remove_non_existing(self) -> int:
+        invalidate = 0
 
+        objects: Dict[str, ImageLabelModel] = {}
+        for image_id, obj in self._datastore.objects.items():
+            if not os.path.exists(os.path.realpath(os.path.join(self._datastore.image_path(), obj.image.path()))):
+                logger.info(f"Removing non existing Image Id: {image_id}")
+                invalidate += 1
+            else:
+                labels: Dict[str, DataModel] = {}
+                for tag, label in obj.labels.items():
+                    if not os.path.exists(
+                        os.path.realpath(os.path.join(self._datastore.label_path(tag), label.path()))
+                    ):
+                        logger.info(f"Removing non existing Label Id: '{label.id}' for '{tag}' for '{image_id}'")
+                        invalidate += 1
+                    else:
+                        labels[tag] = label
+                obj.labels.clear()
+                obj.labels.update(labels)
+                objects[image_id] = obj
+
+        self._datastore.objects.clear()
+        self._datastore.objects.update(objects)
         return invalidate
 
     def _init_from_datastore_file(self, throw_exception=False):
@@ -601,6 +630,7 @@ class LocalDatastore(Datastore):
                     if self._config_ts != ts:
                         logger.debug(f"Reload Datastore; old ts: {self._config_ts}; new ts: {ts}")
                         self._datastore = LocalDatastoreModel.parse_file(self._datastore_config_path)
+                        self._datastore.base_path = self._datastore_path
                         self._config_ts = ts
         except ValueError as e:
             logger.error(f"+++ Failed to load datastore => {e}")
@@ -614,7 +644,7 @@ class LocalDatastore(Datastore):
         logger.debug("+++ Datastore is updated...")
         self._ignore_event_config = True
         with open(self._datastore_config_path, "w") as f:
-            f.write(json.dumps(self._datastore.dict(), indent=2, default=str))
+            f.write(json.dumps(self._datastore.dict(exclude={"base_path"}), indent=2, default=str))
         self._config_ts = os.stat(self._datastore_config_path).st_mtime
 
         if lock:
@@ -622,16 +652,15 @@ class LocalDatastore(Datastore):
 
     def status(self) -> Dict[str, Any]:
         tags: dict = {}
-        for obj in self._datastore.objects:
-            for label in obj.labels:
-                tags[label.tag] = tags.get(label.tag, 0) + 1
+        for obj in self._datastore.objects.values():
+            for tag, _ in obj.labels.items():
+                tags[tag] = tags.get(tag, 0) + 1
 
         return {
             "total": len(self.list_images()),
             "completed": len(self.get_labeled_images()),
             "label_tags": tags,
-            "train": self.datalist(full_path=False),
         }
 
     def json(self):
-        return self._datastore.dict()
+        return self._datastore.dict(exclude={"base_path"})
