+++ plugins/slicer/MONAILabel/MONAILabel.py
@@ -1295,7 +1295,7 @@ class MONAILabelWidget(ScriptedLoadableModuleWidget, VTKObservationMixin):
             logging.info(sample)
             image_id = sample["id"]
             image_file = sample.get("path")
-            image_name = sample.get("name", image_id)[-20:]
+            image_name = sample.get("PatientID", sample.get("name", image_id))[-20:]
             checksum = sample.get("checksum")
             local_exists = image_file and os.path.exists(image_file)
