+++ setup.cfg
@@ -41,6 +41,7 @@ install_requires =
     dicom2nifti==2.3.0
     pydicom-seg==0.2.3
     pynetdicom==1.5.7
+    expiringdict==1.2.1
 
 
 [flake8]
